Zaimplementowano walidatory dla dat oraz enumów, utworzono exception handlera dla status code 400. Zaimplementowano walidacje do sitescontrollera, rozpoczęto implementacje w authcontrollerze. Naprawiano pobieranie drużyn z api dla ligi austriackiej.

+++ src/main/java/com/footballnewsmanager/backend/repositories/NewsTagRepository.java
@@ -3,6 +3,6 @@ package com.footballnewsmanager.backend.repositories;
 import com.footballnewsmanager.backend.models.NewsTag;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface NewsTagRepository extends JpaRepository<NewsTag, Integer> {
+public interface NewsTagRepository extends JpaRepository<NewsTag, Long> {
 
 }
